fix time

+++ src/config/config.cpp
@@ -5,7 +5,7 @@ const String openweathermapServer = "http://api.openweathermap.org/data/2.5/weat
 const String APIKEY = "9881fdc10d1d14339a3a6514d415efa4"; //api openweathermap
 const String ssid = "Home mommy"; // SSID of local network
 const String password = "sapfir1997";   // Password on network
-const uint8_t  timezone = 4;
+const int timezone = 4;
 
 const String meteoserver = "http://meteo-server.herokuapp.com";
 const String routing = "/meteostationData";
